transport/can: migrate legacy CAN to new module.
In preparation for CAN FD support
