+++ uavcan/src/transport/can/tests.rs
@@ -1,8 +1,15 @@
 use alloc::vec;
 
 use crate::time::TestClock;
+use arrayvec::ArrayVec;
+use embedded_time::Clock;
+use num_traits::ToPrimitive;
 
+use super::bitfields::TailByte;
 use super::*;
+use crate::internal::InternalRxFrame;
+use crate::transport::Transport;
+use crate::*;
 
 // I feel I may have gone overboard with these tests, but I'm still getting to grips with
 // testing well so I'm not sure where the boundary should be.
